+++ src/js/components/todo/NewTodoForm.tsx
@@ -4,24 +4,16 @@ import { Todo } from "./type";
 import { Button } from "../parts/Button";
 
 type Props = {
-  setTodoList:  React.Dispatch<React.SetStateAction<Todo[]>>;
+  addTodo: (newTask: string, newDeadline: string, newPerson: string) => void;
 }
 
-export const NewTodoForm = ({ setTodoList }: Props) => {
+export const NewTodoForm = ({ addTodo }: Props) => {
   const [newTask, setNewTask] = React.useState<string>("");
   const [newDeadline, setNewDeadline] = React.useState<string>("");
   const [newPerson, setNewPerson] = React.useState<string>("");
 
   const addNewTodo = () => {
-    setTodoList((prev: Todo[]) => [
-      ...prev,
-      {
-        id: Date.now(),
-        task: newTask,
-        deadline: newDeadline,
-        person: newPerson,
-      },
-    ]);
+    addTodo(newTask, newDeadline, newPerson)
 
     setNewTask("");
     setNewDeadline("");
